editando e estilizando home e repositories

+++ src/pages/Repositories/index.js
@@ -1,16 +1,22 @@
 import React, { useEffect, useState } from 'react';
 import * as S from './styled';
+import { useHistory } from 'react-router-dom';
 
 function Repositories() {
-
+    const history = useHistory();
     const [ repositories, setRepositories ] = useState([]);
 
     useEffect(() => {
         let repositoriesName = localStorage.getItem('repositoriesName');
-        repositoriesName = JSON.parse(repositoriesName);
-        setRepositories(repositoriesName);
-        localStorage.clear();
-    }, []);
+        if (repositoriesName !== null) {
+            repositoriesName = JSON.parse(repositoriesName);
+            setRepositories(repositoriesName);
+            localStorage.clear();
+        }
+        else {
+            history.push('/');
+        }
+    }, [history]);
 
     return (
         <S.Container>
@@ -24,6 +30,7 @@ function Repositories() {
                     }) 
                 }
             </S.List>
+            <S.LinkHome to="/">Voltar</S.LinkHome>
         </S.Container>
     );
 }
